Moved examples to tools
