--- examples/FoamSetup.template/0/p
@@ -1,84 +1,0 @@
-/*--------------------------------*- C++ -*----------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     | Website:  https://openfoam.org
-    \\  /    A nd           | Version:  7
-     \\/     M anipulation  |
-\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       volScalarField;
-    object      p;
-}
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-dimensions      [0 2 -2 0 0 0 0];
-
-internalField   uniform 0.0;
-
-boundaryField
-{
-    movingWall
-    {
-        type            zeroGradient;
-    }
-
-    fixedWall
-    {
-        type            zeroGradient;
-    }
-
-    front
-    {
-        type            zeroGradient;
-    }
-    
-    back
-    {
-        type            zeroGradient;
-    }
-    
-    left
-    {
-        type            zeroGradient;
-    }
-    
-    right
-    {
-        type            zeroGradient;
-    }
-    
-    fronti
-    {
-        type            zeroGradient;
-    }
-    
-    backi
-    {
-        type            zeroGradient;
-    }
-    
-    lefti
-    {
-        type            zeroGradient;
-    }
-    
-    righti
-    {
-        type            zeroGradient;
-    }
-    
-    topi
-    {
-    	type 		 zeroGradient;
-    }
-    
-    buttomi
-    {
-    	type 		 zeroGradient;
-    }
-}
-
-// ************************************************************************* //
