--- examples/FoamSetup.template/system/blockMeshDict
@@ -1,256 +1,0 @@
-/*--------------------------------*- C++ -*----------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     | Website:  https://openfoam.org
-    \\  /    A nd           | Version:  7
-     \\/     M anipulation  |\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       dictionary;
-    object      blockMeshDict;
-}
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-convertToMeters 1.0;
-
-outer 100.0;
-lowerX 35;
-lowerZ 20;
-higherX 65;
-higherZ 50;
-
-vertices
-(
-    (0.0 0.0 0.0) // 0
-    ($lowerX 0.0 0.0)
-    ($higherX 0.0 0.0)
-    ($outer 0.0 0.0)
-    ($outer $outer 0.0) // 4
-    ($higherX $outer 0.0)
-    ($lowerX $outer 0.0)
-    (0.0 $outer 0.0)
-    (0.0 $lowerX 0.0) // 8
-    ($lowerX $lowerX 0.0)
-    ($higherX $lowerX 0.0)
-    ($outer $lowerX 0.0)
-    ($outer $higherX 0.0) // 12
-    ($higherX $higherX 0.0)
-    ($lowerX $higherX 0.0)
-    (0.0 $higherX 0.0)
-    
-    (0.0 0.0 $lowerZ) // 16
-    ($lowerX 0.0 $lowerZ) 
-    ($higherX 0.0 $lowerZ)
-    ($outer 0.0 $lowerZ)
-    ($outer $outer $lowerZ) // 20
-    ($higherX $outer $lowerZ)
-    ($lowerX $outer $lowerZ)
-    (0.0 $outer $lowerZ)
-    (0.0 $lowerX $lowerZ) // 24
-    ($lowerX $lowerX $lowerZ)
-    ($higherX $lowerX $lowerZ)
-    ($outer $lowerX $lowerZ)
-    ($outer $higherX $lowerZ) // 28
-    ($higherX $higherX $lowerZ)
-    ($lowerX $higherX $lowerZ)
-    (0.0 $higherX $lowerZ)
-    
-    (0.0 0.0 $higherZ) // 32
-    ($lowerX 0.0 $higherZ)
-    ($higherX 0.0 $higherZ)
-    ($outer 0.0 $higherZ)
-    ($outer $outer $higherZ) // 36
-    ($higherX $outer $higherZ)
-    ($lowerX $outer $higherZ)
-    (0.0 $outer $higherZ)
-    (0.0 $lowerX $higherZ) // 40
-    ($lowerX $lowerX $higherZ)
-    ($higherX $lowerX $higherZ)
-    ($outer $lowerX $higherZ)
-    ($outer $higherX $higherZ) // 44
-    ($higherX $higherX $higherZ)
-    ($lowerX $higherX $higherZ)
-    (0.0 $higherX $higherZ)
-    
-    (0.0 0.0 $outer) // 48
-    ($lowerX 0.0 $outer)
-    ($higherX 0.0 $outer)
-    ($outer 0.0 $outer)
-    ($outer $outer $outer) // 52
-    ($higherX $outer $outer)
-    ($lowerX $outer $outer)
-    (0.0 $outer $outer)
-    (0.0 $lowerX $outer) // 56
-    ($lowerX $lowerX $outer)
-    ($higherX $lowerX $outer)
-    ($outer $lowerX $outer)
-    ($outer $higherX $outer) // 60
-    ($higherX $higherX $outer)
-    ($lowerX $higherX $outer)
-    (0.0 $higherX $outer)
-);
-
-blocks
-(
-    hex (0 1 9 8 16 17 25 24) (7 7 4) simpleGrading (1 1 1) // 0
-    hex (1 2 10 9 17 18 26 25) (6 7 4) simpleGrading (1 1 1)
-    hex (2 3 11 10 18 19 27 26) (7 7 4) simpleGrading (1 1 1)
-    hex (8 9 14 15 24 25 30 31) (7 6 4) simpleGrading (1 1 1)
-    hex (9 10 13 14 25 26 29 30) (6 6 4) simpleGrading (1 1 1)
-    hex (10 11 12 13 26 27 28 29) (7 6 4) simpleGrading (1 1 1)
-    hex (15 14 6 7 31 30 22 23) (7 7 4) simpleGrading (1 1 1)
-    hex (14 13 5 6 30 29 21 22) (6 7 4) simpleGrading (1 1 1)
-    hex (13 12 4 5 29 28 20 21) (7 7 4) simpleGrading (1 1 1)
-    hex (16 17 25 24 32 33 41 40) (7 7 6) simpleGrading (1 1 1) // 9
-    hex (17 18 26 25 33 34 42 41) (6 7 6) simpleGrading (1 1 1)
-    hex (18 19 27 26 34 35 43 42) (7 7 6) simpleGrading (1 1 1)
-    hex (24 25 30 31 40 41 46 47) (7 6 6) simpleGrading (1 1 1)
-    hex (26 27 28 29 42 43 44 45) (7 6 6) simpleGrading (1 1 1)
-    hex (31 30 22 23 47 46 38 39) (7 7 6) simpleGrading (1 1 1)
-    hex (30 29 21 22 46 45 37 38) (6 7 6) simpleGrading (1 1 1)
-    hex (29 28 20 21 45 44 36 37) (7 7 6) simpleGrading (1 1 1)
-    hex (32 33 41 40 48 49 57 56) (7 7 10) simpleGrading (1 1 1) // 18
-    hex (33 34 42 41 49 50 58 57) (6 7 10) simpleGrading (1 1 1)
-    hex (34 35 43 42 50 51 59 58) (7 7 10) simpleGrading (1 1 1)
-    hex (40 41 46 47 56 57 62 63) (7 6 10) simpleGrading (1 1 1)
-    hex (41 42 45 46 57 58 61 62) (6 6 10) simpleGrading (1 1 1)
-    hex (42 43 44 45 58 59 60 61) (7 6 10) simpleGrading (1 1 1)
-    hex (47 46 38 39 63 62 54 55) (7 7 10) simpleGrading (1 1 1)
-    hex (46 45 37 38 62 61 53 54) (6 7 10) simpleGrading (1 1 1)
-    hex (45 44 36 37 61 60 52 53) (7 7 10) simpleGrading (1 1 1)
-);
-
-edges
-(
-);
-
-boundary
-(
-    movingWall
-    {
-        type wall;
-        faces ((0 8 9 1)
-        	(1 9 10 2)
-        	(2 10 11 3)
-        	(8 15 14 9)
-        	(9 14 13 10)
-        	(10 13 12 11)
-        	(15 7 6 14)
-        	(14 6 5 13)
-        	(13 5 4 12));
-    }
-    
-    fixedWall
-    {
-        type wall;
-        faces ((48 49 57 56)
-		(49 50 58 57)
-		(50 51 59 58)
-		(56 57 62 63)
-		(57 58 61 62)
-		(58 59 60 61)
-		(63 62 54 55)
-		(62 61 53 54)
-		(61 60 52 53));
-    }
-    
-    front
-    {
-        type patch;
-        faces ((0 1 17 16)
-        	(1 2 18 17)
-        	(2 3 19 18)
-        	(16 17 33 32)
-		(17 18 34 33)
-		(18 19 35 34)
-		(32 33 49 48)
-		(33 34 50 49)
-		(34 35 51 50));
-    }
-    
-    back
-    {
-        type patch;
-        faces ((5 21 20 4)
-        	(6 22 21 5)
-        	(7 23 22 6)
-        	(21 37 36 20)
-		(22 38 37 21)
-		(23 39 38 22)
-		(37 53 52 36)
-		(38 54 53 37)
-		(39 55 54 38));
-    }
-    
-    left
-    {
-        type patch;
-        faces ((0 16 24 8)
-        	(8 24 31 15)
-        	(15 31 23 7)
-        	(16 32 40 24)
-		(24 40 47 31)
-		(31 47 39 23)
-		(32 48 56 40)
-		(40 56 63 47)
-		(47 63 55 39));
-    }
-    
-    right
-    {
-        type patch;
-        faces ((3 11 27 19)
-        	(11 12 28 27)
-        	(12 4 20 28)
-       	(19 27 43 35)
-		(27 28 44 43)
-		(28 20 36 44)
-		(35 43 59 51)
-		(43 44 60 59)
-		(44 36 52 60));
-    }
-    
-    fronti
-    {
-    	type patch;
-    	faces ((25 26 42 41));
-    }
-    
-    backi
-    {
-    	type patch;
-    	faces ((29 45 46 30));
-    }
-    
-    lefti
-    {
-    	type patch;
-    	faces ((25 30 46 41));
-    }
-    
-    righti
-    {
-    	type patch;
-    	faces ((29 26 42 45));
-    }
-    
-    topi
-    {
-    	type patch;
-    	faces ((41 42 45 46));
-    }
-    
-    buttomi
-    {
-    	type patch;
-    	faces ((25 26 29 30));
-    }
-);
-
-mergePatchPairs
-(
-);
-
-// ************************************************************************* //
