+++ tools/README.md
@@ -1,0 +1,12 @@
+# MaMiCo Tools
++ __debugging/compare_sim_output.py__ Compares the output generated by two simulation runs qualitatively.
+Output files are paired by their suffix (`<timestep>.<extension>`) and must have matching prefixes (before first `_`).
+Why? Allows for easy comparison of the simulation output between different implementations
+(e.g. verify correctness of simulation after non-trivial refactoring).
++ __debugging/hash_sim_output.py__ Creates a list of hashes and output file suffixes (`..._<timestep>.<extension>`).
+Why? Allows for easy comparison of the simulation output between different implementations
+(e.g. verify correctness of simulation after non-trivial refactoring).
++ __plotting/plot-couette-profiles.py__ Simple but versatile script to plot couette flow profiles based on the CSV output.
++ __plotting/plot-pint-LBCouette.py__ Used to create Couette flow profile plots (average X-vel over simulation time) for a time-parallel simulation.
+Plots several PinT iterations in a single plot, can be used to show convergence.
++ __mamico-cfg__ Utility script to generate basic couette.xml configurations (and additional files) for MaMiCo.
